requirements.txt
Flask
opencv-python
Pillow
gunicorn
numpy